Security tweaks.
Removed the early exits from DecryptAndValidateAuthenticationCookie
to reduce the chance of a potential timing attack. Also removed the
separator in the cookie and replaced it with a calcuation of the
base64 length of the hmac size (which is always the same for the same
hashing algorithm). Thanks to @Blowdart for letting me pick his brain :-)