Updated NDjango view engine and added extends demo.
Refactored the ndjango view engine a bit to prevent the
'double location' issue. Now means we have to construct
the TemplateProviderManager per request, but that is
quicker than double location and perf is pretty much
as good as it was.

Added an 'extends' route/views to the main demo project
and it seems to still do what it is supposed to :-)