Added custom build providers for razor views
Added Nancy.ViewEngines.Razor.BuildProviders project and updated the
Razor nuget so that it will automatically install the build providers
into the bin folder of the project it's installed into. It will hook
up a post-build event hook to make sure the dll is always in the bin
at runtime.