Added more SDHP to the favicon overriding
The loading of the favicon is moved to FavIconStartup. The startup
will first try to locate a favicon.ico or favicon.png file anywhere
in the project (Using the current IRootProvider). If that fails then
it will fallback to use the embedded Nancy.favicon.ico file in the
main Nancy dll