Disposing body stream at end of request
The NancyContext will now dispose the Request, when done, which in turn
will dispose the Body stream (RequestStream instance). If the stream has
been switched to a new stream (memory or file) then that stream will also
be disposed. If the stream has been switched to the filesystem then the
temporary file is attempted to be deleted as well, so we do not leave
any garbage files around.