Updated the DotLiquid engine
* Added unit tests for the LiquidNancyFileSystem implementation and
also updated it to use template locations

* Removed tests from DotLiquidViewEngine that were testing the
LiquidNancyFileSystem through the DotLiquidViewEngine implementation

* Introduced IFileSystemFactory and DefaultFileSystemFactory which is
used to create a filesystem for the provided context. Updated the
DotLiquidViewEngine to take a ctor dependency on the interface and
updated the Initialize method to use the factory

* Added DotLiquidViewEngineStartup to make sure that the
DefaultFileSystemFactory is registered in the container

* Added unit test to ensure that the factory is used in the Initialize
method of the engine

* Added support to LiquidNancyFileSystem so the name of the template
can .liquid extension, but still works without it. Also made it support
both forward- and backslashed in the location path. Made all matching
logic case-insensitive

* Removed convention that stripped underscores from the template name. It
made no sense to have it