Spike for negotiating response format
Added a /negotiated route to the Nancy.Demo.Hosting.Aspnet project
that can be used to get a negotiated response. Currently it can
return either json and xml (the two ISerializer implementations that
are shipped with Nancy).

Also adds the Vary: Accept header if there are more than one serializer
that can be used for the request.