Changed auto registration behavior
The DisableAutoRegistration function on the ConfigurableBootstrapper
has been replaced by EnableAutoRegistration. This means that with this
patch, the auto-registration will be disabled by default instead of
enabled. The reason for this are primarily for performance but also
to help make tests be more implicit in setting up the context in
which they execute.