Added trailing slashes (\) to post-build steps
Added trailing slashes (\) to post-build xcopy steps to fix an issue when building under
TFS whereby XCOPY tries to prompt the user to choose to copy into the bin folder, or copy as a file called 'bin'. This was causing the TFS build process to hang forever, requireing the build to be stopped.