Added conventon based assembly loading
The ConfigurableBootstrapper will now attempt to locate the
name of the test assembly. Once it has a hold of the name it
will see if it can strip "test", "tests", "unittests", "specs"
or "specifications" from the name.

It will assume that the resulting name is the name of the
assembly that is under test and will identify all the referenced
assemblies and load them into the application domain.

The list of suffixes can be modified by changing the static
field TestAssemblySuffixes on ConfigurableBoostrapper.

Fixes issue #421