Added support for route descriptions
RouteDescriptions got a Description property that is set by the
RouteCache at application startup. The cache uses the new
IRouteDescriptionProvider interface to get the description.

The default implementation of the IRouteDescriptionProvider looks
for the descriptions in resource files. If you create a resource
file, with the same name as the module that has the route(s) you
want to add descriptions for, and use the route path as the key
then it will automatically be picked up. Note that you have to set
your resource as "no code behind" in order to use paths as resource
keys

You can swap out the IRouteDescriptionProvider, for your own one,
in the internal configuration. This enables you to store the
descriptions anywhere!