Added support for Context in SSVE
You can now access the context by using @Context and it works the
same way as @Model. In addition the conditional @If and @IfNot
as well as the @Each operator now supports that you deine the
model source using

  @If.Model...
  @IfNot.Model...
  @If.Context...
  @IfNot.Context...
  @Each.Model...
  @Each.Context...

If the model source is not included it will default to use the model
as the source.