Added new Nancy.Owin project
Includes basic extenion method for wiring up with IAppBuilder
and also the Nancy OWIN host, with the main entry point renamed
to Invoke to fit in with the standard conventions.

Moving forwards the signature of INancyEngine may change to become
more compatible with OWIN (returning Task), but all that should be
hidden behind the extension methods provided in here.

Conflicts:
	src/Nancy.sln