Fixed #908
Accept header now always set if we go through negotiation. Initially
considered only setting it if the negotiator could match more than
one type (either by having a wildcard or multiple entries), but there's
still a chance that a processor may match in some weird and wonderful
way behind the scenes so seems sensible to always add it if we have
negotiated the response.