Fixed HEAD and OPTIONS requests and mspec tests
Mspec test was picking up the wrong module.

HEAD and OPTIONS requests weren't implemented in the new route
resolver so added tests and code for them.

In two minds whether to cache the OPTIONS results, because it does
multiple route resolutions, but the trie is fast enough that
hopefully it won't be necessary.