Removed empty view location convention from error
A view location convention returns empty at times and those will
never be used to try and locate the view. If a view cannot be found
then we display an error page with the conventions that was used and
it looks a bit strange with all the empty entries.

This change excludes all empty convention results from the error page
and only display those conventions that was actually used to attempt
to locate the view.