Fixed #1090
Only seek the source stream on switch if seeking is available.
Technically we could get into a position where we lose some of the body
contents, but as we can't check the current position on the self host
stream that isn't seekable there's no much we can do inside the
RequestStream class.