Removed dependency on System.Drawing
We had a dependency on System.Drawing because we used it in the
FavIconApplicationStartup class to load a favicon to disk and
convert it to a byte array (with the help of an ImageConverter).

This made it impossible to use Nancy with Mono to target Android.
I've replaced the System.Drawing dependency with the use of
File.ReadAllBytes instead.

Hope I didn't break anything ;)