Adjust TryParse method to allow more convertions
Adjust TryParse method to allow more convertions.
. Convertions like Int64 to Int32 were ignored in the previous version,
since there was no ending else statement.
Now the last rule is to try a type convertion.
. Convertions to a nullable type were also ignored (like Int64 to
Int64?), now there is a test to check if a nullable type convertion is
possible.

Note that if the convertion fails any exception will be treated and the
default value returned.

Some test cases were created for these situations.

 