Update how to build instructions with missing prerequisites
The rake build is not compatible with the latest version of albacore, it's only compatible with v1.0.0.rc2.
If you don't have Visual Studio 2010 installed then you'll need to install the Visual Studio 2010 shell
in order to build projects that reference Microsoft.WebApplication.targets. The Windows SDK is required
because Microsoft.Common.targets reference AL.exe using the SdkToolsPath.