Added ISerializerFactory + implementation
The idea is that this what you will take a dependency on, instead of
IEnumerable<ISerializer>, and it will give you the serializer that
matches the provided MediaRange.

The default implementation will take the default serializers into
account and only return them if no other match could be made.

If there are multiple matched (not counting the defaults) then it will
throw an InvalidOperationException with a detailed error message, telling
you about the types of the implementaions that it matched.

I have also updated the ISerializer.CanSerialize and Serialize methods
to accept a MediaRange instance, instead of string, for the contentType
parameter. This works fine since it can implicitly be case to/from a
string.