Fixes the assemby loading into the temp AppDomain
As specified by MSDN, AppDomain.Load() is just a convenience method for
Assembly.Load(). This means that the previous behavior loaded all the
assemblies into both the inspection AppDomain and AppDomain.Current, when
the code intent was clear that this was not desired.

However, if the assembly is passed between AppDomains, because it is not
serializable, it would be loaded as well into both AppDomains, therefore
the solution is to make a proxy prober which loads the assembly in the
inspection domain and locally determines if the desired reference exists.